refactor(py): genkit.core.schemas -> genkit.core.schema_types since we also have a schema module
Rename the module from schemas to schema_types to avoid collisions
with another module called 'schema' that will be ported soon.

CHANGELOG:
- [ ] Update the name of the module.
- [ ] Update references to the name.
- [ ] Update ruff configuration to avoid ignoring lint in this module.
