+++ py/bin/generate_schema_types
@@ -6,7 +6,7 @@
 set -euo pipefail
 
 TOP_DIR=$(git rev-parse --show-toplevel)
-SCHEMA_FILE="${TOP_DIR}/py/packages/genkit/src/genkit/core/schemas.py"
+SCHEMA_FILE="${TOP_DIR}/py/packages/genkit/src/genkit/core/schema_types.py"
 
 # Generate types using configuration from pyproject.toml
 uv run --directory "${TOP_DIR}/py" datamodel-codegen
@@ -15,7 +15,7 @@ uv run --directory "${TOP_DIR}/py" datamodel-codegen
 #sed -i '' '/^class Model(RootModel\[Any\]):$/,/^    root: Any$/d' "${SCHEMA_FILE}"
 
 # Sanitize the generated schema.
-python3 "${TOP_DIR}/py/bin/sanitize_schemas.py" "${SCHEMA_FILE}"
+python3 "${TOP_DIR}/py/bin/sanitize_schema_types.py" "${SCHEMA_FILE}"
 
 # Checks and formatting.
 uv run --directory "${TOP_DIR}/py" \
