+++ py/bin/sanitize_schema_types.py
@@ -4,9 +4,9 @@
 # SPDX-License-Identifier: Apache-2.0
 
 
-"""Standalone convenience script used to massage the schemas.py.
+"""Standalone convenience script used to massage the schema_types.py.
 
-The `py/packages/genkit/src/genkit/core/schemas.py` file is generated by
+The `py/packages/genkit/src/genkit/core/schema_types.py` file is generated by
 datamodel-codegen. However, since the tool doesn't currently provide options to
 generate exactly the kind of code we need, we use this convenience script to
 parse the Python source code, walk the AST, modify it to include the bits we
@@ -21,8 +21,8 @@ Transformations applied:
   Pythonic API but serialize to camelCase in order to be compatible with
   runtimes.
 - We add a license header
-- We add a header indicating that this file has been generated by a code generator
-  pass.
+- We add a header indicating that this file has been generated by a code
+  generator pass.
 - We add the ability to use forward references.
 """
 
@@ -83,7 +83,8 @@ class ModelConfigRemover(ast.NodeTransformer):
         return False
 
     def visit_ClassDef(self, node: ast.ClassDef) -> ast.ClassDef:  # noqa: N802
-        """Visit class definitions and handle model_config based on class type."""
+        """Visit class definitions and handle model_config based on class
+        type."""
         if self.is_rootmodel_class(node):
             # Filter out model_config assignments for RootModel classes
             new_body = []
@@ -111,7 +112,8 @@ class ModelConfigRemover(ast.NodeTransformer):
                     node.body.insert(0, self.create_model_config())
                     self.modified = True
                 else:
-                    # Update existing model_config to include populate_by_name=True
+                    # Update existing model_config to include
+                    # populate_by_name=True
                     new_body = []
                     for item in node.body:
                         if isinstance(item, ast.Assign):
